feat: implement custom hooks for file upload, image refresh, and upload events in UploadSprout component
